netdevice: disable netdevice use of romfile option
Some cases, VMs want a nic but do not want it to be used during boot
(no PXE) which can be acheived by passing 'romfile=' (no value set)
This prevents the built-in romfile from being loaded and the nic will
not be presented to the BIOS as a bootable device.

NetDevice.ROMFile = DisabledNetDeviceROMFile

will emit the 'romfile=' in the command line preventing nic from
being used during BIOS boot.