feat: add support for vvfat blockdev
QEMU's vvfat driver can export a directory as a VFAT filsystem into
the guest.  In particular, we'd like to be able to set the volume
label so that cloud-init can detect the fs as cloud-config.  The
qemu -drive cli doesn't expose the label property but it is available
via the -blockdev command.  Here we special case VVFAT in blockdev
until an overhaul is done converting all -drive use into -blockdev.

Signed-off-by: Ryan Harper <rharper@woxford.com>